refactor(application): extract jobApplicationService
