+++ src/runtime/appContext.ts
@@ -42,6 +42,7 @@ import { EnvApplicationServiceImpl } from '../application/impl/envApplicationSer
 import { ProfileApplicationServiceImpl } from '../application/impl/profileApplicationServiceImpl.js';
 import { DoctorApplicationServiceImpl } from '../application/impl/doctorApplicationServiceImpl.js';
 import { InitApplicationServiceImpl } from '../application/impl/initApplicationServiceImpl.js';
+import { JobApplicationServiceImpl } from '../application/impl/jobApplicationServiceImpl.js';
 import { loadConfigSync } from '../utils/config.js';
 import { resolveWorkspaceDir } from '../utils/instance.js';
 import { createDefaultLogger, createSilentLogger, setDefaultLogger } from '../utils/logger.js';
@@ -67,6 +68,7 @@ import type { EnvApplicationService } from '../application/envApplicationService
 import type { ProfileApplicationService } from '../application/profileApplicationService.js';
 import type { DoctorApplicationService } from '../application/doctorApplicationService.js';
 import type { InitApplicationService } from '../application/initApplicationService.js';
+import type { JobApplicationService } from '../application/jobApplicationService.js';
 import type { AiConfig } from '../utils/types/index.js';
 
 export interface AppContext {
@@ -90,6 +92,7 @@ export interface AppContext {
   profileApp: ProfileApplicationService;
   doctorApp: DoctorApplicationService;
   initApp: InitApplicationService;
+  jobApp: JobApplicationService;
   // config
   defaultAiConfig: AiConfig;
 }
@@ -146,6 +149,7 @@ function wireContext(
   const profileApp = new ProfileApplicationServiceImpl();
   const doctorApp = new DoctorApplicationServiceImpl(profileRepository);
   const initApp = new InitApplicationServiceImpl();
+  const jobApp = new JobApplicationServiceImpl(jobRepo, companyRepo);
 
   return {
     jobRepository: jobRepo,
@@ -165,6 +169,7 @@ function wireContext(
     profileApp,
     doctorApp,
     initApp,
+    jobApp,
     defaultAiConfig,
   };
 }
